requirements.txt
asgiref==3.7.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs
bcrypt
beautifulsoup4==4.9.3
click
dnspython==2.7.0
environs==14.1.0
fastapi==0.115.6
fastapi_cors==0.0.6
Flask==2.0.2
Flask-Cors==5.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
h11==0.14.0
idna
itsdangerous
Jinja2
jsonschema
motor==3.6.0
pymongo==4.9.2
pydantic
python-dateutil
python-dotenv
requests==2.26.0
starlette==0.41.3
SQLAlchemy
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==2.0.2
WTForms==3.2.1
